The DCComic class checks if the given comic (or panel) already exists.

+++ dinosaurcomics/DCComic.py
@@ -19,8 +19,11 @@ class DCComic():
         return path.join(self._comics_dir, '{0}.png'.format(self._comic_id))
 
     def save_comic(self):
-        r = get(BS(get(COMIC_URL.format(self._comic_id), headers=hdrs).text).select('.comic')[0]['src'], stream=True)
         fn = path.join(self._comics_dir, '{0}.png'.format(self._comic_id))
+        if path.exists(fn):
+            return
+
+        r = get(BS(get(COMIC_URL.format(self._comic_id), headers=hdrs).text).select('.comic')[0]['src'], stream=True)
         if r.status_code == 200:
             with open(fn, 'wb') as f:
                 for chunk in r.iter_content(1024):
@@ -28,6 +31,10 @@ class DCComic():
 
     def save_panel(self, p):
         fn = path.splitext(self.comic_path)
+        panel_fn = '{0}_{1}{2}'.format(fn[0], p, fn[1])
+        if path.exists(panel_fn):
+            return
+
         with Image(filename=self.comic_path) as img:
             img.crop(**panels[p])
-            img.save(filename='{0}_{1}{2}'.format(fn[0], p, fn[1]))
+            img.save(filename=panel_fn)
